Move rootComponents out of pages add resolve aliass
