+++ assets/src/components/root/issues.js
@@ -1,8 +1,8 @@
-import comment from '../comment.vue'
-import issueComment from '../issueComment.vue'
-import openClose from '../openClose.vue'
-import switchEditApi from '../switchEditApi.vue'
-import inlineEdit from '../inlineEdit.vue'
+import comment from '@/comment.vue'
+import issueComment from '@/issueComment.vue'
+import openClose from '@/openClose.vue'
+import switchEditApi from '@/switchEditApi.vue'
+import inlineEdit from '@/inlineEdit.vue'
 
 export default {
     el: '#vue-container',
