+++ assets/src/components/root/tagPage.js
@@ -1,0 +1,8 @@
+import switchEditComponent from '@/switchEdit.vue'
+
+export default {
+    el: '#vue-container',
+    components: {
+        'switch-edit': switchEditComponent
+    }
+}
