+++ assets/src/pages/issue.js
@@ -1,7 +1,7 @@
 import { initCore } from '../lib/init.js'
 
 import Vue from 'vue'
-import rootComponent from '../components/root/issue.js'
+import rootComponent from '@/root/issues.js'
 
 
 initCore()
