+++ assets/src/pages/tagPage.js
@@ -1,19 +1,10 @@
 import { initCore } from '../lib/init.js'
-import switchEditComponent from '../components/switchEdit.vue'
 
 import Vue from 'vue'
+import rootComponent from '@/root/tagPage.js'
 
-import '../styles/taggle.css'
-import '../styles/aristotle.dashboard.less'
+import 'src/styles/taggle.css'
+import 'src/styles/aristotle.dashboard.less'
 
 initCore()
-
-// Export root component for testing
-export var rootComponent = {
-    el: '#vue-container',
-    components: {
-        'switch-edit': switchEditComponent
-    }
-}
-
 new Vue(rootComponent)
